9.8 20:14

+++ app/src/main/java/com/fika/fika_project/retrofit/DTOClass.kt
@@ -22,5 +22,5 @@ data class testerCode(
 )
 data class CourseEditDTO(
     val courseTitle : String?,
-    val spotIdList : ArrayList<Int>
+    val spotIdList : ArrayList<Long>
 )
